Added async validator to check if classes of stream aren't empty
